refactor: move SpaceFilling implementation into cpp files
